# docs/features/internal-api.openapi.yaml
openapi: 3.1.0
info:
  title: MySpeedPuzzling Internal Admin API
  version: "1.0.0"
  description: |
    Admin-only API for triggering privileged operations from outside the shell
    (Claude Code, curl, phone). Completely separate from the public
    `/api/v1/*` OAuth2 API — uses a single static bearer token.

    Closed-by-default: if `INTERNAL_API_TOKEN` env var is empty,
    all endpoints return 401.

    Not wired into Swagger UI at `/api/docs` on purpose. Render this spec
    manually in Swagger Editor / Redoc / an IDE plugin.
servers:
  - url: "{scheme}://{host}"
    variables:
      scheme:
        default: https
      host:
        default: localhost:8080
security:
  - InternalApiBearer: []
paths:
...
  /internal-api/competitions:
    get:
      summary: List or search competitions
      operationId: listCompetitions
      description: |
        Every competition - approved, pending, rejected or a draft, standalone or an
        edition of a series - newest first (by the event's date, undated ones by when
        they were added).
      parameters:
        - name: q
          in: query
          required: false
          description: Part of the name, slug or shortcut of the competition or of its series (accents and letter case ignored).
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: The approval state (`approved` / `pending` / `rejected` - an approved draft is approved and not pending), or `draft` - the competition or its series is a draft.
          schema:
            type: string
            enum: [all, approved, pending, rejected, draft]
            default: all
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: One page of competitions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionList"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Create a competition
      operationId: createCompetition
      description: |
        A standalone competition (past or upcoming), added by the reviewer player,
        validated by the rules of the web form (`CompetitionFormData`). With
        `"approve": true` it is approved right away. Under an approved organization
        (`organizationId`) it is approved at once anyway - the reviewer player is an
        admin (`OrganizationApprovalPolicy`). `"draft": true` creates it as a draft:
        only its team sees it until it is published. Rounds and puzzles follow
        through their own endpoints. No e-mail is sent.

        Requires `INTERNAL_API_REVIEWER_PLAYER_ID` to be configured.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              unevaluatedProperties: false
              allOf:
                - $ref: "#/components/schemas/CompetitionFields"
                - type: object
                  required: [name]
                  properties:
                    approve:
                      type: boolean
                      default: false
                    organizationId:
                      type: [string, "null"]
                      format: uuid
                      description: An existing organization (400 otherwise) - the one-time event is under it.
                    draft:
                      type: boolean
                      default: false
      responses:
        "201":
          description: Created - the competition as `GET /internal-api/competitions/{competitionId}` shows it
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionDetail"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  # One path item for both: GET also takes a slug in place of the id, PATCH only the id
...
components:
  securitySchemes:
    InternalApiBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Static token from the `INTERNAL_API_TOKEN` env var.
...
  schemas:
    CompetitionFields:
      type: object
      # No additionalProperties here - it would not see the create bodies' own fields (allOf); the request
      # bodies close themselves with unevaluatedProperties: false (an unknown field is a 400)
      description: |
        Validated by the web form's rules: name required; an in-person event
        (`isOnline` false) needs `location`, `dateFrom` and `dateTo` (an edition of
        a series needs neither); `dateTo` not before `dateFrom`, at most 30 days apart;
        links must be URLs. `null` or `""` clears a field. An unknown field is a 400.
      properties:
        name:
          type: string
          maxLength: 250
        shortcut:
          type: [string, "null"]
          maxLength: 250
        description:
          type: [string, "null"]
        location:
          type: [string, "null"]
          maxLength: 250
        locationCountryCode:
          type: [string, "null"]
          description: ISO 3166-1 alpha-2, any letter case (stored lower case).
          example: cz
        dateFrom:
          type: [string, "null"]
          description: ISO 8601 day (`2026-11-14`); a date-time counts by its day.
        dateTo:
          type: [string, "null"]
        link:
          type: [string, "null"]
          format: uri
          description: The event's website.
        registrationLink:
          type: [string, "null"]
          format: uri
        resultsLink:
          type: [string, "null"]
          format: uri
        isOnline:
          type: boolean
          description: An online event has no place - true clears `location`, like the web form. Its dates stay (optional for an online event - empty = ongoing).
        eligibility:
          type: [string, "null"]
          maxLength: 120
          description: '"Who can enter" ("Residents of the state", "21+") - an edition without its own shows its series''.'
        slug:
          type: string
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          description: Explicit URL slug - unique (409 otherwise). Left out, create generates it from the name and update keeps it. A change breaks the old URLs (no redirect); it cannot be cleared.
        maintainerIds:
          type: array
          description: Player ids who may manage the event - replaces the list, `[]` empties it, `null` or left out keeps it.
          items:
            type: string
            format: uuid
    Competition:
      type: object
      properties:
        competitionId: { type: string, format: uuid }
        name: { type: string }
        slug: { type: [string, "null"] }
        shortcut: { type: [string, "null"] }
        description: { type: [string, "null"] }
        location: { type: [string, "null"] }
        locationCountryCode: { type: [string, "null"] }
        dateFrom: { type: [string, "null"], example: "2026-11-14" }
        dateTo: { type: [string, "null"] }
        link: { type: [string, "null"] }
        registrationLink: { type: [string, "null"] }
        resultsLink: { type: [string, "null"] }
        isOnline: { type: boolean }
        isRecurring:
          type: boolean
          description: True for an edition of a series (`series`) - recurring events are series of editions.
        series:
          type: [object, "null"]
          properties:
            seriesId: { type: string, format: uuid }
            name: { type: string }
            slug: { type: [string, "null"] }
            organizationId:
              type: [string, "null"]
              format: uuid
              description: The series' organization - an edition's organization is always its series'.
            draft: { type: boolean, description: The series is a draft (it hides this edition too) }
        organizationId:
          type: [string, "null"]
          format: uuid
          description: A one-time event's own organization - always null for an edition (see `series.organizationId`).
        organization:
          type: [object, "null"]
          properties:
            organizationId: { type: string, format: uuid }
            name: { type: string }
            slug: { type: string }
        eligibility: { type: [string, "null"], description: '"Who can enter" - its own; an edition without one shows its series''' }
        logo: { type: [string, "null"] }
        tagId: { type: [string, "null"] }
        tagName: { type: [string, "null"] }
        status:
          type: string
          enum: [approved, pending, rejected]
          description: The approval state (`IsCompetitionPubliclyVisible::SQL_APPROVED`) - an approved draft is `approved`; an edition is approved through its series.
        draft: { type: boolean, description: The competition's own draft flag }
        hiddenAsDraft: { type: boolean, description: It or its series is a draft }
        approvedAt: { type: [string, "null"], format: date-time }
        approvedByPlayerId: { type: [string, "null"] }
        rejectedAt: { type: [string, "null"], format: date-time }
        rejectionReason: { type: [string, "null"] }
        publiclyVisible:
          type: boolean
          description: The whole visibility rule (`IsCompetitionPubliclyVisible`) - approval and drafts.
        createdAt: { type: [string, "null"], format: date-time }
        addedByPlayerId: { type: [string, "null"] }
        addedByPlayerName: { type: [string, "null"] }
        roundsCount: { type: integer }
        resultsCount: { type: integer, description: Solving times linked to the competition (suspicious ones included) }
        resultsWithoutRoundCount: { type: integer, description: Of those, the ones in no round (their puzzle is in none of its rounds of their category) }
        participantsCount: { type: integer, description: Participants who joined (not removed) }
    CompetitionList:
      type: object
      required: [total, limit, offset, competitions]
      properties:
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        competitions:
          type: array
          items:
            $ref: "#/components/schemas/Competition"
    CompetitionDetail:
      allOf:
        - $ref: "#/components/schemas/Competition"
        - type: object
          properties:
            maintainers:
              type: array
              items:
                type: object
                properties:
                  playerId: { type: string, format: uuid }
                  name: { type: [string, "null"] }
                  code: { type: string }
            rounds:
              type: array
              items:
                $ref: "#/components/schemas/Round"
            puzzles:
              type: array
              description: The competition's own (tagged) puzzles.
              items:
                $ref: "#/components/schemas/AdminPuzzle"
...
    Round:
      type: object
      properties:
        roundId: { type: string, format: uuid }
        competitionId: { type: string, format: uuid }
        slug: { type: [string, "null"] }
        name: { type: string }
        category: { type: string, enum: [solo, duo, team] }
        startsAt: { type: string, format: date-time, example: "2026-11-14T09:00:00+00:00" }
        timezone: { type: string, example: Europe/Prague, description: The round's own zone }
        minutesLimit: { type: integer }
        revealDelayMinutes:
          type: integer
          minimum: 0
          maximum: 240
          example: 10
          description: Minutes after the start when the round's secret puzzles with an automatic reveal come out
        teamSize:
          type: [integer, "null"]
          example: 4
          description: Team rounds - how many people a team is expected to have (null = not set)
        badgeBackgroundColor: { type: [string, "null"] }
        badgeTextColor: { type: [string, "null"] }
        resultsLink: { type: [string, "null"] }
        resultsCount:
          type: integer
          description: Solving times that belong to the round.
        puzzles:
          type: array
          items:
            allOf:
              - $ref: "#/components/schemas/AdminPuzzle"
              - type: object
                properties:
                  roundPuzzleId: { type: string, format: uuid }
                  hideUntilRoundStarts: { type: boolean, description: Secret until its reveal }
                  hideMode: { type: [string, "null"], enum: [image_only, entirely, null] }
                  revealMode:
                    type: string
                    enum: [automatic, scheduled, manual]
                    description: automatic = the round's `revealDelayMinutes` after it starts (follows its start and delay); scheduled = the organiser's own moment; manual = when the organiser clicks "Reveal now"
                  revealsAt:
                    type: [string, "null"]
                    format: date-time
                    description: The reveal moment (UTC) of a secret row - null when not secret, or manual and not revealed yet
                  hidesEverywhere:
                    type: boolean
                    description: The round keeps the puzzle secret on the whole site, not only on its event pages
...
    AdminPuzzle:
      type: object
      properties:
        puzzleId: { type: string, format: uuid }
        name: { type: string }
        piecesCount: { type: integer }
        manufacturerId: { type: [string, "null"] }
        manufacturerName: { type: [string, "null"] }
        ean: { type: [string, "null"], description: "Comma-separated when the puzzle has several" }
        identificationNumber: { type: [string, "null"], description: Brand code(s) }
        approved: { type: boolean }
        hiddenUntil: { type: [string, "null"], format: date-time, description: "The site-wide hide of the name (a secret competition puzzle, a placeholder)" }
        imageHiddenUntil: { type: [string, "null"], format: date-time, description: The site-wide hide of the picture }
...
    Error:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string
        errors:
          type: object
          description: Invalid fields (field => what is wrong) - only for a request with invalid fields.
          additionalProperties:
            type: string
        revealedPuzzles:
          type: array
          description: |
            Only present when a round change, a round delete or a puzzle removal would
            reveal secret puzzles earlier than planned. It lists what
            `"confirmReveal": true` would let out. A round PATCH that moves the
            automatic reveal earlier lists them at the new moment, or right away when
            that moment is already over.
          items:
            type: object
            required: [puzzleId, name, rightAway, revealsAt, previousRevealsAt, scope, revealedEverywhere, stillHiddenElsewhereUntil]
            properties:
              puzzleId: { type: string, format: uuid }
              name: { type: string }
              rightAway:
                type: boolean
                description: true = out the moment the request is applied (`revealsAt` is null)
              revealsAt:
                type: [string, "null"]
                format: date-time
                description: When it would come out (UTC) - the new automatic reveal moment; null = right away
              previousRevealsAt:
                type: [string, "null"]
                format: date-time
                description: |
                  The moment it moves from (UTC): for a round PATCH, the round's automatic
                  reveal as it is now. Null for a removal or a deleted round (the puzzle
                  leaves the round).
              scope:
                type: string
                enum: [everywhere, name_everywhere, event]
                description: |
                  How far it would come out:
                  - `everywhere`: everything the round hid, on the whole site.
                  - `name_everywhere`: its name on the whole site; its picture stays
                    hidden elsewhere until `stillHiddenElsewhereUntil`.
                  - `event`: on this event only; another round keeps it hidden
                    elsewhere until `stillHiddenElsewhereUntil` - or, when that is null,
                    it was public elsewhere all along (a public catalogue puzzle the
                    round keeps secret on its event pages only).
              revealedEverywhere:
                type: boolean
                description: true when `scope` is `everywhere` (it existed before `scope` and stays)
              stillHiddenElsewhereUntil:
                type: [string, "null"]
                format: date-time
                description: |
                  For `name_everywhere` and `event`, until when the rest stays hidden
                  elsewhere (UTC); null for `everywhere`, and for `event` when the puzzle
                  was public elsewhere all along. `9999-12-31T00:00:00+00:00` = until a
                  manual reveal ("Reveal now" in another round) is made.
...
  responses:
...
    Unauthorized:
      description: Missing, wrong, or unconfigured bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InvalidRequest:
      description: Malformed JSON, a missing or invalid field, or no reviewer configured
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The request conflicts with the current state (taken slug, round invariant, results in a round, secret puzzles a round change would reveal, a hidden puzzle attached unhidden, shared tag, already approved, known EAN, an organization on an edition, a move that cannot be made, cannot go back to draft, an organization that is not empty)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Unknown feature request, merge request, brand, puzzle, competition, round, series, organization or player ID
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"